refactor(asr): extract processing/reconstruction and QA helpers
Move process_asr + the Riemannian processing backends to _reconstruction.py and
compute_asr_qa_metrics / compute_asr_rejection_mask to _qa.py (the latter
type-imports ASR under TYPE_CHECKING to avoid a runtime cycle). core.py
re-exports everything via __all__; no behavior change. core.py is now ~1045
lines with only the ASR estimator class left to move.
